Add SQL-Stmt Definition to TableDataObject trait
