+++ sdl-core/src/main/scala/io/smartdatalake/workflow/dataobject/JdbcTableDataObject.scala
@@ -90,10 +90,10 @@ import scala.util.Try
  */
 case class JdbcTableDataObject(override val id: DataObjectId,
                                createSql: Option[String] = None,
-                               preReadSql: Option[String] = None,
-                               postReadSql: Option[String] = None,
-                               preWriteSql: Option[String] = None,
-                               postWriteSql: Option[String] = None,
+                               override val preReadSql: Option[String] = None,
+                               override val postReadSql: Option[String] = None,
+                               override val preWriteSql: Option[String] = None,
+                               override val postWriteSql: Option[String] = None,
                                override val schemaMin: Option[GenericSchema] = None,
                                override var table: Table,
                                override val constraints: Seq[Constraint] = Seq(),
@@ -402,23 +402,8 @@ case class JdbcTableDataObject(override val id: DataObjectId,
     )
   }
 
-  override def preRead(partitionValues: Seq[PartitionValues])(implicit context: ActionPipelineContext): Unit = {
-    super.preRead(partitionValues)
-    prepareAndExecSql(preReadSql, Some("preReadSql"), partitionValues)
-  }
-  override def postRead(partitionValues: Seq[PartitionValues])(implicit context: ActionPipelineContext): Unit = {
-    super.postRead(partitionValues)
-    prepareAndExecSql(postReadSql, Some("postReadSql"), partitionValues)
-  }
-  override def preWrite(implicit context: ActionPipelineContext): Unit = {
-    super.preWrite
-    prepareAndExecSql(preWriteSql, Some("preWriteSql"), Seq()) // no partition values here...
-  }
-  override def postWrite(partitionValues: Seq[PartitionValues])(implicit context: ActionPipelineContext): Unit = {
-    super.postWrite(partitionValues)
-    prepareAndExecSql(postWriteSql, Some("postWriteSql"), partitionValues)
-  }
-  private def prepareAndExecSql(sqlOpt: Option[String], configName: Option[String], partitionValues: Seq[PartitionValues])(implicit context: ActionPipelineContext): Unit = {
+
+  def prepareAndExecSql(sqlOpt: Option[String], configName: Option[String], partitionValues: Seq[PartitionValues])(implicit context: ActionPipelineContext): Unit = {
     sqlOpt.foreach { sql =>
       val data = DefaultExpressionData.from(context, partitionValues)
       val preparedSql = SparkExpressionUtil.substitute(id, configName, sql, data)
